Rename to normalized_env
